+++ NEWS.md
@@ -1,3 +1,11 @@
+# interpM
+
+## New features
+
+* Adiciona metodo `rbind.curvacolina`. Normalmente so tera uso interno em conjunto com os 
+  interpoladores multiplos. Deve ser notado que o objeto retornado remove quaisquer linhas 
+  duplicadas que possam resultar da combinacao
+
 # newtri
 
 ## New features