+++ NEWS.md
@@ -1,4 +1,4 @@
-# modvaz
+# curvacolina 1.9.1
 
 ## New features